added video card component
